refactor(batch): refactor executor micro benchmark framework
